LPS-75598 Rename UADEntity*Impl.java to BaseUADEntity*.java and update package
