+++ modules/apps/foundation/user-associated-data/user-associated-data-api/bnd.bnd
@@ -3,5 +3,4 @@ Bundle-SymbolicName: com.liferay.user.associated.data.api
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.user.associated.data.model,\
-	com.liferay.user.associated.data.model.impl,\
 	com.liferay.user.associated.data.service.registry
