--- modules/apps/foundation/user-associated-data/user-associated-data-api/src/main/resources/com/liferay/user/associated/data/model/impl/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
